% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{add_feature_ids}
\alias{add_feature_ids}
\title{Add `id` column to a dataframe}
\usage{
add_feature_ids(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
The same dataframe but with an additional `id` column
in first place which contains the rownames.
}
\description{
Add `id` column to a dataframe
}
