% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.step.hybrid.R
\name{align_recovered_batch_tables}
\alias{align_recovered_batch_tables}
\title{Internal function: Align batch tables returned by feature_recovery.}
\usage{
align_recovered_batch_tables(
  recovered_tables,
  batch_names,
  batch_align_mz_tol,
  batch_align_rt_tol,
  min_occurrence
)
}
\arguments{
\item{recovered_tables}{A list of tibbles containing joined metadata, intensity and rt tibbles produced by recover_weaker_signals.}

\item{batch_names}{A list of batch names - treated as sample names.}

\item{batch_align_mz_tol}{The m/z tolerance for batch alignment.}

\item{batch_align_rt_tol}{The retention time tolerance for batch alignment.}

\item{min_occurrence}{A feature has to show up in at least this number of profiles to be included in the final result.}
}
\value{
A list of aligned metadata, intensity and rt tibbles across batches.
}
\description{
Combines individual batch tibbles into one list object containing aligned metadata, intensity and rt tibbles from all samples.
}
