% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{as_feature_sample_table}
\alias{as_feature_sample_table}
\title{Read the metadata table, retention time data matrix and intensity data matrix
and combine them into a single table}
\usage{
as_feature_sample_table(metadata, rt_crosstab, int_crosstab)
}
\arguments{
\item{metadata}{Tibble Feature metadata table with information concerning the peaks.}

\item{rt_crosstab}{Tibble Data matrix with features on rows and samples on columns holding rt data.}

\item{int_crosstab}{Tibble Data matrix with features on rows and samples on columns holding intensity data.}
}
\value{
Tibble A merged table containing all information.
}
\description{
Read the metadata table, retention time data matrix and intensity data matrix
and combine them into a single table
}
