% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{augment_known_table}
\alias{augment_known_table}
\title{Add newly detected aligned features to a known features table.}
\usage{
augment_known_table(
  aligned,
  known_table,
  match_tol_ppm,
  mz_tol_relative,
  rt_tol_relative,
  new_feature_min_count
)
}
\arguments{
\item{aligned}{A list object with three tibble tables: metadata, intensity, and rt.}

\item{known_table}{A table of known/previously detected peaks.}

\item{match_tol_ppm}{The ppm tolerance to match identified features to known metabolites/features.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
the tolerance level based on the data.}

\item{new_feature_min_count}{The number of profiles a new feature must be present for it to be added to the database.}
}
\value{
Known table with novel features.
}
\description{
Add newly detected aligned features to a known features table.
}
