% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{augment_with_known_features}
\alias{augment_with_known_features}
\title{Add newly detected features to a known features table.}
\usage{
augment_with_known_features(
  aligned,
  known_table,
  match_tol_ppm,
  mz_tol_relative,
  rt_tol_relative
)
}
\arguments{
\item{aligned}{A list object with three tibble tables: metadata, intensity, and rt.}

\item{known_table}{A table of known/previously detected peaks.}

\item{match_tol_ppm}{The ppm tolerance to match identified features to known metabolites/features.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
the tolerance level based on the data.}
}
\value{
Known table with novel features.
}
\description{
Add newly detected features to a known features table.
}
