% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{bigauss.esti.EM}
\alias{bigauss.esti.EM}
\title{Compute bi-gaussian parameters using Expectation-Maximization algorithm}
\usage{
bigauss.esti.EM(
  t,
  x,
  max.iter = 50,
  epsilon = 0.005,
  do.plot = FALSE,
  sigma_ratio_lim = c(0.3, 1)
)
}
\arguments{
\item{t}{A vector of numerical values (rt).}

\item{x}{A vector of numerical values (intensities).}

\item{max.iter}{Maximum number of iterations.}

\item{epsilon}{Threshold for continuing the iteration}

\item{sigma_ratio_lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard deviation 
and the right-standard deviation of the bi-Gaussian function.}
}
\value{
A vector with length 4. The items are as follows going from first to last:
\itemize{
  \item mean of gaussian curve
  \item standard deviation at the left side of the gaussian curve
  \item standard deviation at the right side of the gaussian curve
  \item estimated total signal strength (total area of the estimated normal curve)
}
}
\description{
Computes bi-gaussian parameters using an iterative method - Expectation-Maximization algorithm.
The algorithm iteratively updates the parameters until convergence or until the maximum number of iterations is reached.
}
