% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{bigauss.esti}
\alias{bigauss.esti}
\title{Estimate the parameters of Bi-Gaussian curve.}
\usage{
bigauss.esti(x, y, power = 1, do.plot = FALSE, sigma.ratio.lim = c(0.3, 3))
}
\arguments{
\item{x}{Vector of RTs that lay in the same RT cluster.}

\item{y}{Intensities that belong to x.}

\item{power}{The power parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture model in an EIC.}

\item{sigma.ratio.lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard deviation
and the right-standard deviation of the bi-Gaussian function used to fit the data.}
}
\value{
A vector with length 4. The items are as follows going from first to last:
\itemize{
  \item mean of gaussian curve
  \item standard deviation at the left side of the gaussian curve
  \item standard deviation at the right side of the gaussian curve
  \item estimated total signal strength (total area of the estimated normal curve)
}
}
\description{
Estimate the parameters of Bi-Gaussian curve.
}
