% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{bigauss.mix}
\alias{bigauss.mix}
\title{Bi-gaussian mixture model estimation}
\usage{
bigauss.mix(
  rt_profile,
  moment_power = 1,
  do.plot = FALSE,
  sigma_ratio_lim = c(0.1, 10),
  bw = c(15, 30, 60),
  eliminate = 0.05,
  max.iter = 25,
  peak_estim_method,
  BIC_factor = 2
)
}
\arguments{
\item{rt_profile}{Dataframe that stores RTs and intensities of features.}

\item{moment_power}{The parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture model in an EIC.}

\item{sigma_ratio_lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard deviation
and the right-standard deviation of the bi-Gaussian function used to fit the data.}

\item{bw}{Bandwidth vector to use in the kernel smoother.}

\item{eliminate}{When a component accounts for a proportion of intensities less than this value, the component will be ignored.}

\item{max.iter}{Maximum number of iterations when executing the E step.}

\item{peak_estim_method}{The estimation method for the bi-Gaussian peak model. Two possible values: moment and EM.}

\item{BIC_factor}{The factor that is multiplied on the number of parameters to modify the BIC criterion. If larger than 1,
models with more peaks are penalized more.}
}
\description{
Estimates the optimal bi-gaussian parameters using the EM method. It accepts two internal computation of parameters for "moment"
and "EM" model input options.
}
