% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.characterize.R
\name{characterize}
\alias{characterize}
\title{Internal function: Update feature characteristics in the known feature table row.}
\usage{
characterize(existing_row, n, m, metadata_row, rt_row, intensity_row)
}
\arguments{
\item{existing_row}{A vector representing an existing row in the known feature table with 18 elements:
\itemize{
  \item [1-5]: Chemical identifiers and metadata
  \item [6]: m.z - Mean m/z value
  \item [7]: Number_profiles_processed - Total number of profiles processed
  \item [8]: Percent_found - Proportion of profiles where feature was found
  \item [9]: mz_min - Minimum m/z value observed
  \item [10]: mz_max - Maximum m/z value observed
  \item [11]: RT_mean - Mean retention time
  \item [12]: RT_sd - Standard deviation of retention time
  \item [13]: RT_min - Minimum retention time
  \item [14]: RT_max - Maximum retention time
  \item [15]: int_mean(log) - Mean log-intensity
  \item [16]: int_sd(log) - Standard deviation of log-intensity
  \item [17]: int_min(log) - Minimum log-intensity
  \item [18]: int_max(log) - Maximum log-intensity
}}

\item{n}{The number of times the feature was found in previous experiments.}

\item{m}{The number of times the feature was found in the current experiment.}

\item{metadata_row}{A row from the aligned feature metadata table containing mz, mzmin, and mzmax.}

\item{rt_row}{A vector of retention times for the feature across samples in the current experiment.}

\item{intensity_row}{A vector of log-transformed intensities for the feature across samples.}
}
\value{
A vector with the updated statistics for all 18 elements of the known feature table row.
}
\description{
This function updates a single row in the known feature table by merging existing statistics
with new data from the current experiment. It updates the number of profiles processed,
percent found, m/z range, retention time statistics, and intensity statistics.
}
