% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{check_files}
\alias{check_files}
\title{Check files whether they exist.}
\usage{
check_files(filenames)
}
\arguments{
\item{filenames}{list of filenames Filenames to check whether they exist.}
}
\description{
Check files whether they exist.
}
