% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{clean_data_matrix}
\alias{clean_data_matrix}
\title{Replace NA values by zero, relocate 'sample_names' column to the very beginning and convert to a tibble}
\usage{
clean_data_matrix(x, sample_names)
}
\arguments{
\item{x}{A dataframe}

\item{sample_names}{List of sample names.}
}
\value{
Cleaned tibble.
}
\description{
Replace NA values by zero, relocate 'sample_names' column to the very beginning and convert to a tibble
}
