% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{comb}
\alias{comb}
\title{Combines the output (i.e. metadata, intensity and RT) from different clusters to one respective tibble.}
\usage{
comb(x, ...)
}
\value{
Tibbles combining the output (metadata, intensity and RT respectively) from different clusters.
}
\description{
Combines the output (i.e. metadata, intensity and RT) from different clusters to one respective tibble.
}
