% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_all_times}
\alias{compute_all_times}
\title{Compute all time values from base curve.}
\usage{
compute_all_times(base_curve)
}
\arguments{
\item{base_curve}{Basis curve}
}
\description{
Compute all time values from base curve.
}
