% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_boundaries}
\alias{compute_boundaries}
\title{Compute bounds of area using given peak and mass valley points.}
\usage{
compute_boundaries(valley_points, peak)
}
\arguments{
\item{valley_points}{vector values of valley points defining clusters.}

\item{peak}{double Value of the peak for which to get the valley bounds.}
}
\value{
Returns a list object with the following objects in it:
\itemize{
  \item lower - double - The value of the lower bound valley point.
  \item upper - double - The value of the upper bound valley point.
}
}
\description{
The lower bound is the mass of the valley the closest but smaller than peak
and the upper bound is the mass of the valley the closest but higher than
the peak.
}
