% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_bounds}
\alias{compute_bounds}
\title{Computes initial and final bounds of set of values.}
\usage{
compute_bounds(x, sigma_ratio_lim)
}
\arguments{
\item{x}{Cumulative intensity values.}

\item{sigma_ratio_lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard deviation.
and the right-standard deviation of the bi-Gaussian function used to fit the data.}
}
\value{
Returns a list with bounds with following items:
\itemize{
  \item start - start bound
  \item end - end bound
}
}
\description{
Computes initial and final bounds of set of values.
}
