% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.tol.time.R
\name{compute_breaks_3}
\alias{compute_breaks_3}
\title{Compute indices of mass differences greater than min_mz_tol.}
\usage{
compute_breaks_3(values, tolerance = NA)
}
\arguments{
\item{tolerance}{tolerance value.}

\item{mz}{mz values of all peaks in all profiles in the study.}
}
\value{
breaks Integer indices of mass differences to use.
}
\description{
Compute indices of mass differences greater than min_mz_tol.
}
