% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_chromatographic_profile}
\alias{compute_chromatographic_profile}
\title{Find base.curve RTs that lay within RT range of the whole feature table and append the intensities to these RTs.}
\usage{
compute_chromatographic_profile(profile, base.curve)
}
\arguments{
\item{profile}{Profile table with shape number-of-features*4. The table contains following columns:
\itemize{
  \item mz - float - mass-to-charge ratio of feature
  \item rt - float - retention time of features
  \item intensity - float - intensity of features
  \item group_number - integer - group number assigned to each feature based on their rt similarity
}}

\item{base.curve}{Matrix that contains rts of feature in the same rt cluster.}
}
\value{
dataframe with two columns
}
\description{
Find base.curve RTs that lay within RT range of the whole feature table and append the intensities to these RTs.
}
