% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{compute_clusters_simple}
\alias{compute_clusters_simple}
\title{Compute clusters using simple grouping based on numeric thresholds.}
\usage{
compute_clusters_simple(feature_tables, sample_names, mz_tol_ppm, rt_tol)
}
\arguments{
\item{feature_tables}{list of tibbles List of feature tables coming from all samples.}

\item{sample_names}{list of strings Sample names of the feature tables used to distinguish the samples.}

\item{mz_tol_ppm}{float Relative tolerance for mz grouping in parts per million.}

\item{rt_tol}{float Tolerance in retention time dimension [seconds].}
}
\value{
list of tibbles Feature tables passed initially with additional columns indicating the 
mz and rt groups as well as the combined cluster index.
}
\description{
Compute clusters using simple grouping based on numeric thresholds.
}
