% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{compute_clusters_simple_sd}
\alias{compute_clusters_simple_sd}
\title{Compute clusters using simple grouping based on numeric thresholds.}
\usage{
compute_clusters_simple_sd(
  feature_tables,
  sample_names,
  mz_tol_ppm,
  rt_tol,
  sd_ratio_tol
)
}
\arguments{
\item{feature_tables}{list of tibbles feature tables coming from all samples.}

\item{sample_names}{list of strings Sample names of the feature tables used to distinguish the samples.}

\item{mz_tol_ppm}{float Relative tolerance for mz grouping in parts per million.}

\item{rt_tol}{float Tolerance in retention time dimension [seconds].}

\item{sd_ratio_tol}{float sd ratio tolerance to use for grouping features.}
}
\value{
list of tibbles Feature tables passed initially with additional columns indicating the 
mz and rt groups as well as the combined cluster index.
}
\description{
Features are first grouped in mz dimension based on the tolerance.
First, the absolute tolerance is computed for each feature, then a new group is started
once the difference between consecutive features is above this threshold.
The same process is then repeated for the retention time dimension.
The individual indices are then combined into a single index in the `cluster` columns.
}
