% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{compute_comb}
\alias{compute_comb}
\title{Combine template and sample features}
\usage{
compute_comb(template_features, features)
}
\arguments{
\item{template_features}{Tibble Template feature table (mz, rt, cluster, sample_id).}

\item{features}{Tibble Sample feature table (mz, rt, cluster, sample_id).}
}
\value{
Tibble Combined feature table (rbind).
}
\description{
Combine template and sample features
}
