% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{compute_corrected_features}
\alias{compute_corrected_features}
\title{Correct the rt in feature table based on paired feature rts and differences.}
\usage{
compute_corrected_features(features, template_rt, delta_rt)
}
\arguments{
\item{features}{Tibble The feature table for which to correct rts.}

\item{template_rt}{List of floats Template retention times for the paired features.}

\item{delta_rt}{List of floats Differences between the paired rts.}
}
\value{
Tibble A table with corrected retention times.
}
\description{
Correct the rt in feature table based on paired feature rts and differences.
}
