% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_curr_rec_with_enough_peaks}
\alias{compute_curr_rec_with_enough_peaks}
\title{Compute the rectangle around recovered features given that enough peaks are present.}
\usage{
compute_curr_rec_with_enough_peaks(
  mz,
  peak,
  valleys,
  features,
  aver_diff,
  times,
  delta_rt
)
}
\arguments{
\item{mz}{Mz value of the feature.}

\item{peak}{Peak around which to detect the new feature.}

\item{valleys}{Valley points to compute the boundary region.}

\item{features}{tibble Tibble with `rt` and `intensities` column.}

\item{aver_diff}{float Average retention time difference.}

\item{times}{vector Raw retention time values from raw data file.}

\item{delta_rt}{vector Differences between consecutive retention time values (diff(times)).}
}
\value{
list Triplet of mz, label and intensity for the feature.
}
\description{
Compute the rectangle around recovered features given that enough peaks are present.
}
