% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_delta_rt}
\alias{compute_delta_rt}
\title{Compute custom smoothed h-method derivative of function.}
\usage{
compute_delta_rt(times)
}
\arguments{
\item{times}{Retention time values.}
}
\value{
Differences between time values.
}
\description{
The function adds an extrapolated element in the end and a 0 element in the front,
then computes the midpoints between#' neighbouring elements and then uses the `diff`
function to compute the changes in rt between points.
}
