% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive.bin.R
\name{compute_densities}
\alias{compute_densities}
\title{Compute densities for given masses.}
\usage{
compute_densities(
  masses,
  mz_tol,
  intensity_weighted,
  intensity,
  bw_func,
  n = 512
)
}
\arguments{
\item{masses}{A numeric vector of mass values.}

\item{mz_tol}{The mass-to-charge ratio tolerance.}

\item{intensity_weighted}{A logical value indicating whether to weight the densities by intensities.}

\item{intensity}{A numeric vector of intensity values corresponding to the masses.}

\item{bw_func}{A function to compute the bandwidth based on the masses.}

\item{n}{The number of equally spaced points at which the density is to be estimated. Default is 512.}
}
\value{
A density object representing the estimated density of the masses.
}
\description{
Compute the density estimation for a set of masses using kernel density estimation.
It allows for intensity-weighted densities and custom bandwidth functions.
}
