% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_dx}
\alias{compute_dx}
\title{Compute difference between neighbouring elements of a vector and optionally apply a mask such that the maximum difference is no higher than 4-fold minimum difference.}
\usage{
compute_dx(x, apply_mask = TRUE)
}
\arguments{
\item{x}{- float - a vector of numerical values.}

\item{apply_mask}{- boolean - whether to apply threshold mask to the output vector.}
}
\value{
Returns vector of numeric differences between neighbouring values.
}
\description{
Compute difference between neighbouring elements of a vector and optionally apply a mask such that the maximum difference is no higher than 4-fold minimum difference.
}
