% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_e_step}
\alias{compute_e_step}
\title{Computes the expectation step of the EM method.}
\usage{
compute_e_step(rt_profile, miu, s1, s2, delta)
}
\arguments{
\item{rt_profile}{A matrix with two columns: "base.curve" (rt) and "intensity".}

\item{s1}{Parameter computed by the initiation step.}

\item{s2}{Parameter computed by the initiation step.}

\item{delta}{Parameter computed by the initiation step.}

\item{m}{A vector of sorted RT-peak values at which the kernel estimate was computed.}
}
\description{
Computes the expectation step of the EM method.
}
