% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_gaussian_peak_shape}
\alias{compute_gaussian_peak_shape}
\title{Compute parameters of chromatographic peak shape if peaks are considered to be gaussian}
\usage{
compute_gaussian_peak_shape(
  rt_profile,
  power,
  bw,
  component.eliminate,
  BIC.factor,
  aver_diff
)
}
\arguments{
\item{rt_profile}{A matrix with two columns: "base.curve" (rt) and "intensity".}

\item{power}{The power parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture model in an EIC.}

\item{bw}{Bandwidth vector to use in the kernel smoother.}

\item{component.eliminate}{When a component accounts for a proportion of intensities less than this value, the component will be ignored.}

\item{BIC.factor}{The factor that is multiplied on the number of parameters to modify the BIC criterion. If larger than 1,
models with more peaks are penalized more.}

\item{aver_diff}{Average retention time difference across RTs of all features.}
}
\value{
Returns a single-row vector or a table object with the following items/columns:
\itemize{
  \item miu - float - mean value of the gaussian curve
  \item sigma - float - standard deviation of the gaussian curve
  \item sigma - float - standard deviation of the gaussian curve
  \item scale - float - estimated total signal strength (total area of the estimated normal curve)
}
}
\description{
Compute parameters of chromatographic peak shape if peaks are considered to be gaussian
}
