% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_initiation_params}
\alias{compute_initiation_params}
\title{Calculates the three initial bi-gaussian parameters (sd1, sd2, and scaling factor)}
\usage{
compute_initiation_params(rt_profile, pks, vlys, dx)
}
\arguments{
\item{rt_profile}{A matrix with two columns: "base.curve" (rt) and "intensity".}

\item{pks}{A vector of sorted RT-peak values at which the kernel estimate was computed.}

\item{vlys}{A vector of sorted RT-valley values at which the kernel estimate was computed.}

\item{dx}{Difference between neighbouring RT values with step 2.}
}
\value{
A matrix. The items are as follows going from first to last:
\itemize{
  \item s1: standard deviation at the left side of the gaussian curve
  \item s2: standard deviation at the right side of the gaussian curve
  \item delta: estimated total signal strength (total area of the estimated normal curve)
}
probably a matrix is better though
}
\description{
Calculates the three initial bi-gaussian parameters (sd1, sd2, and scaling factor)
}
