% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.step.hybrid.R
\name{compute_intensity_medians}
\alias{compute_intensity_medians}
\title{Internal function: Compute median intensities for features.}
\usage{
compute_intensity_medians(feature_table)
}
\arguments{
\item{feature_table}{A tibble containing feature data with a sample_intensity column.}
}
\value{
The feature table with an additional median_intensity column.
}
\description{
Calculates the median intensity for each feature across all samples in which it appears.
The median is computed by grouping features by their m/z and retention time.
}
