% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_mass_density}
\alias{compute_mass_density}
\title{Compute the density function of mz values.}
\usage{
compute_mass_density(features, bandwidth, intensity_weighted, n = 512)
}
\arguments{
\item{bandwidth}{Bandwidth to use to compute the kernel density.}

\item{intensity_weighted}{Whether to use intensity weighting or not.}

\item{mz}{Mass values to compute the density over.}

\item{intensities}{Intensities of the peaks at mz values.
Only used if intensity_weighted == TRUE.}
}
\value{
\link[stats]{density} object containing the densities.
}
\description{
The function takes the mz values and uses \link[stats]{density} to
compute the local density, optionally using intensity based weighting.
}
