% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive.bin.R
\name{compute_mass_values}
\alias{compute_mass_values}
\title{Compute mass values based on kernel density estimation.}
\usage{
compute_mass_values(mz_tol, masses, intensity_binned, intensity_weighted)
}
\arguments{
\item{mz_tol}{The mass-to-charge ratio tolerance.}

\item{masses}{A numeric vector of mass values.}

\item{intensity_binned}{A numeric vector of binned intensity values corresponding to the masses.}

\item{intensity_weighted}{A logical value indicating whether to weight the densities by intensities.}
}
\value{
A numeric vector of mass values corresponding to the valleys in the density plot.
}
\description{
This function computes the mass values by performing kernel density estimation on the given masses.
It identifies the valleys in the density plot to determine the mass values.
}
