% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mz_tolerance.R
\name{compute_min_mz_tolerance}
\alias{compute_min_mz_tolerance}
\title{Compute minimum mz tolerance to use.}
\usage{
compute_min_mz_tolerance(mz, mz_tol_relative, mz_tol_absolute)
}
\arguments{
\item{mz}{vector Mz values to use.}

\item{mz_tol_relative}{float Relative mz tolerance to use with the mz values.
This forms a sort of weighted tolerance.}

\item{mz_tol_absolute}{float Absolute tolerance to use independent from the mz values.}
}
\value{
float Minimum tolerance values to use.
}
\description{
Compute the minimum mz tolerance based on the relative
tolerance and the mz values and the absolute tolerance.
Uses midpoints between mz values for the weighting.
}
