% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mz_sd}
\alias{compute_mz_sd}
\title{Compute standard deviation of m/z values for feature groups.}
\usage{
compute_mz_sd(feature_groups)
}
\arguments{
\item{feature_groups}{A list of data frames, where each data frame represents a group of features with m/z values.}
}
\value{
A numeric vector of standard deviations of m/z values for each feature group.
}
\description{
This function computes the standard deviation of m/z values for each group of features.
}
