% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_peaks_and_valleys}
\alias{compute_peaks_and_valleys}
\title{Compute peaks and valleys of density function.}
\usage{
compute_peaks_and_valleys(dens)
}
\arguments{
\item{density}{stats::density Density object for which to compute peaks
and valleys.}
}
\value{
Returns a list object with the following objects in it:
\itemize{
  \item pks - vector - The data points at which the density peaks.
  \item vlys - vector - The points in the data where the density is low
                        (forming a valley in the function).
}
}
\description{
Given a density function, the turn points are computed and
the peaks and valleys in the original data (points with highest
and lowest density) are returned.
}
