% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_pks_vlys_rt}
\alias{compute_pks_vlys_rt}
\title{Compute peaks and valleys which have at least `recover_min_count` peaks.}
\usage{
compute_pks_vlys_rt(features, times, bandwidth, target_rt, recover_min_count)
}
\arguments{
\item{features}{tibble Features with `mz`, `rt` and `intensity`.}

\item{times}{vector Retention time values from the raw data file.}

\item{bandwidth}{float Bandwidth to use in smoothing.}

\item{target_rt}{float Retention time at which to recover the intensity.}

\item{recover_min_count}{int Minimum number of peaks required in the area to recover the signal.}
}
\value{
Returns a list object with the following objects in it:
\itemize{
  \item pks - vector - The data points at which the density peaks with at least `recover_min_count` peaks between the valley points.
  \item vlys - vector - The points in the data where the density is low
                        (forming a valley in the function).
}
}
\description{
Compute peaks and valleys which have at least `recover_min_count` peaks.
}
