% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_rectangle}
\alias{compute_rectangle}
\title{Compute rectangle around feature with `aligned_feature_mz` and `target_rt` for recovery.}
\usage{
compute_rectangle(
  data_table,
  aligned_feature_mz,
  breaks,
  custom_mz_tol,
  orig_mz_tol,
  use_intensity_weighting,
  recover_min_count,
  target_rt,
  custom_rt_tol,
  times,
  delta_rt,
  aver_diff,
  bandwidth,
  min_bandwidth,
  max_bandwidth
)
}
\arguments{
\item{data_table}{tibble Feature table with `mz`, `rt` and `intensities` column.}

\item{aligned_feature_mz}{float Mz value of feature in aligned feature table.}

\item{breaks}{vector Integer boundaries of clusters in mz values.}

\item{custom_mz_tol}{float Custom mz tolerance for the feature.}

\item{orig_mz_tol}{float Flat original mz tolerance to use.}

\item{use_intensity_weighting}{bool Whether to use intensity weighting.}

\item{recover_min_count}{int Minimum number of peaks required in the area to recover the signal.}

\item{target_rt}{float Target retention time value.}

\item{custom_rt_tol}{float Custom chromatographic tolerance to use.}

\item{times}{vector Raw retention time values from raw data file.}

\item{delta_rt}{vector Differences between consecutive retention time values (diff(times)).}

\item{aver_diff}{float Average retention time difference.}

\item{bandwidth}{float Bandwidth to use in smoothing.}

\item{min_bandwidth}{float Minimum bandwidth to use.}

\item{max_bandwidth}{float Maximum bandwidth to use.}
}
\value{
tibble Tibble with `mz`, `rt` and `intensities` columns.
}
\description{
Compute rectangle around feature with `aligned_feature_mz` and `target_rt` for recovery.
}
