% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.ridge.R
\name{compute_rt_intervals_indices}
\alias{compute_rt_intervals_indices}
\title{Computes indices of retention time intervals of a selected retention time range.}
\usage{
compute_rt_intervals_indices(rt, rt_sel)
}
\arguments{
\item{rt}{full retention time vector.}

\item{rt_sel}{Selected retention time range vector.}
}
\value{
A list object:
\itemize{
  \item over_rt - indices of upper retention time interval
  \item under_rt - indices of lower retention time interval
  \item within_rt - indices of intermediate retention time interval
}
}
\description{
Computes indices of retention time intervals of a selected retention time range.
}
