% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.tol.time.R
\name{compute_rt_tol_relative}
\alias{compute_rt_tol_relative}
\title{Compute rt relative tolerance to use.}
\usage{
compute_rt_tol_relative(
  breaks,
  max.num.segments,
  aver.bin.size,
  number_of_samples,
  rt,
  min.bins,
  max.bins,
  do.plot = FALSE
)
}
\arguments{
\item{max.num.segments}{the maximum number of segments.}

\item{aver.bin.size}{The average bin size to determine the number of equally spaced points in the kernel density estimation.}

\item{number_of_samples}{The number of spectra in this analysis.}

\item{rt}{Retention time of all peaks in all profiles in the study.}

\item{min.bins}{The minimum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too few observations are present.}

\item{max.bins}{The maximum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too many observations are present.}

\item{do.plot}{Indicates whether plot should be drawn.}
}
\value{
rt_tol_relative The elution time tolerance.
}
\description{
Compute the elution time tolerance based on the kernel density estimation.
It plots the fitting function if set to TRUE.
}
