% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{compute_sel}
\alias{compute_sel}
\title{Select features to use for retention time alignment}
\usage{
compute_sel(combined)
}
\arguments{
\item{combined}{Tibble Table with (mz, rt, cluster, sample_id).}
}
\value{
List of bool Returns list of bools with TRUE at each index where this condition is met.
}
\description{
This function selects features present in both the sample
feature table and template feature table given they have the same cluster,
are adjacent in the combined table.
}
