% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_start_bound}
\alias{compute_start_bound}
\title{Computes initial bound of set of values.}
\usage{
compute_start_bound(x, left_sigma_ratio_lim)
}
\arguments{
\item{x}{Cumulative intensity values.}

\item{left_sigma_ratio_lim}{Left-standard deviation of the bi-Gaussian function.}
}
\value{
Returns end bound.
}
\description{
Computes initial bound of set of values.
}
