% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_target_times}
\alias{compute_target_times}
\title{Compute target times for regions of interest for recovery.}
\usage{
compute_target_times(
  aligned_rts,
  original_sample_rts,
  adjusted_sample_rts,
  min_common_times = 4
)
}
\arguments{
\item{aligned_rts}{vector Aligned retention time values.}

\item{original_sample_rts}{vector Original feature retention time values before correction.}

\item{adjusted_sample_rts}{vector Feature retention time values after time correction.}
}
\description{
Compute the individual target times for the features to be recovered in the sample.
Spline interpolation using \link[splines]{interpSpline} is used to map from adjusted times
back into the original times. The function requires `x` to be distinct, hence the filtering
to only include rt values that occurr only once in both lists.
}
