% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{compute_template}
\alias{compute_template}
\title{Select the template feature table.}
\usage{
compute_template(extracted_features)
}
\arguments{
\item{extracted_features}{List of tables Tables from which to select the template.}
}
\value{
Tibble Template feature table.
}
\description{
The current implementation selects the table with the most features as the template.
}
