% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{compute_template_adjusted_rt}
\alias{compute_template_adjusted_rt}
\title{Create two column table with paired sample and template retention times.}
\usage{
compute_template_adjusted_rt(combined, sel, j)
}
\arguments{
\item{combined}{Tibble Table with features from sample and template.}

\item{sel}{list of bools List of bools indiciating which features to pair.
See 'compute_sel'.}

\item{j}{string Template sample_id.}
}
\description{
Create two column table with paired sample and template retention times.
}
