% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_filter.R
\name{compute_uniq_grp}
\alias{compute_uniq_grp}
\title{Computes unique groups}
\usage{
compute_uniq_grp(profile, min_count_run, min_pres)
}
\arguments{
\item{profile}{The matrix containing m/z, retention time, intensity, and EIC label as columns.}

\item{min_count_run}{filter parameter.}

\item{min_pres}{Run filter parameter. The minimum proportion of presence in the time period for a series of signals grouped
by m/z to be considered a peak.}
}
\value{
unique_grp.
}
\description{
Computes unique groups
}
