% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_feature_tables}
\alias{concatenate_feature_tables}
\title{Concatenate multiple feature lists and add the sample id (origin of feature) as additional column.}
\usage{
concatenate_feature_tables(features, sample_names)
}
\arguments{
\item{features}{list List of tibbles containing extracted feature tables.}
}
\description{
Concatenate multiple feature lists and add the sample id (origin of feature) as additional column.
}
