% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{correct_time}
\alias{correct_time}
\title{Function to perform retention time correction}
\usage{
correct_time(this.feature, template_features)
}
\arguments{
\item{this.feature}{Tibble Feature table for which to correct rt.}

\item{template_features}{Tibble Template feature table to use for correction.}
}
\value{
Tibble this.feature table with corrected rt values.
}
\description{
Function to perform retention time correction
}
