% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{correct_time_v2}
\alias{correct_time_v2}
\title{Rewritten version of 'correct_time'}
\usage{
correct_time_v2(features, template)
}
\arguments{
\item{features}{Tibble Table with features to correct.}

\item{template}{Tibble Template feature table to use for correction.}
}
\value{
Tibble Corrected feature table.
}
\description{
This function uses dplyr to do the same as
'correct_time', just with less code. Most functions used in the original
function are replaced with simple data transformations.
}
