% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{count_peaks}
\alias{count_peaks}
\title{Count the number of peaks in all valleys}
\usage{
count_peaks(roi, times)
}
\arguments{
\item{roi}{list Named list with vectors `pks` and `vlys`.}

\item{times}{vector Retention time values}
}
\value{
vector Numbers of peaks within each region defined by a peak and the two valley points.
}
\description{
For each peak in ROI, count the peaks between the valley points.
}
