% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_aligned_feature_table}
\alias{create_aligned_feature_table}
\title{Align peaks from spectra into a feature table.}
\usage{
create_aligned_feature_table(
  features_table,
  min_occurrence,
  sample_names,
  rt_tol_relative,
  mz_tol_relative,
  cluster = 4
)
}
\arguments{
\item{features_table}{A list object. Each component is a matrix which is the output from compute_clusters().}

\item{min_occurrence}{A feature has to show up in at least this number of profiles to be included in the final result.}

\item{sample_names}{list List of sample names.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which
allows the program to search for the tolerance level based on the data.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the
program to search for the tolerance level based on the data. This value is expressed as the
percentage of the m/z value. This value, multiplied by the m/z value, becomes the cutoff level.}

\item{cluster}{The number of CPU cores to be used}
}
\value{
A tibble with three tables containing aligned metadata, intensities an RTs.
}
\description{
Align peaks from spectra into a feature table.
}
