% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_aligned_feature_table_simple}
\alias{create_aligned_feature_table_simple}
\title{Simple version of create_aligned_feature_table for testing, grouping by cluster IDs}
\usage{
create_aligned_feature_table_simple(
  features_table,
  min_occurrence,
  sample_names,
  batch = FALSE
)
}
\arguments{
\item{features_table}{A list object. Each component is a matrix which is the output from compute_clusters().}

\item{min_occurrence}{A feature has to show up in at least this number of profiles to be included in the final result.}

\item{sample_names}{List of sample names.}

\item{batch}{A flag to signify if the function works with batch data.}
}
\value{
A list of 3 tibbles containing aligned metadata, intensities an RTs.
}
\description{
Simple version of create_aligned_feature_table for testing, grouping by cluster IDs
}
