% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_features_from_cluster}
\alias{create_features_from_cluster}
\title{Group the mz and RT for particular cluster.}
\usage{
create_features_from_cluster(
  features,
  mz_tol_relative,
  rt_tol_relative,
  min_occurrence,
  sample_names
)
}
\arguments{
\item{features}{The features table subsetted for a particular cluster.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment.}

\item{min_occurrence}{A minimal number of profiles a feature has to be present in.}

\item{sample_names}{A list of sample names.}
}
\value{
A list containing 3 tibbles: metadata, intensities and RTs.
}
\description{
Group the mz and RT for particular cluster.
}
