% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_features_from_cluster_simple}
\alias{create_features_from_cluster_simple}
\title{Simple version of create_features_from_cluster for testing}
\usage{
create_features_from_cluster_simple(
  features,
  min_occurrence,
  sample_names,
  batch = FALSE
)
}
\arguments{
\item{features}{The features table subsetted for a particular cluster.}

\item{min_occurrence}{A minimal number of profiles a feature has to be present in}

\item{sample_names}{A list of sample names.}

\item{batch}{A flag to signify if the function works with batch data.}
}
\value{
A list containing 3 tibbles: metadata, intensities and RT
}
\description{
Simple version of create_features_from_cluster for testing
}
