% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_intensity_row}
\alias{create_intensity_row}
\title{Compute summed area for each sample}
\usage{
create_intensity_row(sample_grouped)
}
\arguments{
\item{sample_grouped}{A dataframe with grouped mz and RT values for a particular cluster.}
}
\value{
Summed area for each sample.
}
\description{
Compute summed area for each sample
}
