% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_intensity_row_tsh}
\alias{create_intensity_row_tsh}
\title{Create intensity row for batch data}
\usage{
create_intensity_row_tsh(sample_grouped)
}
\arguments{
\item{sample_grouped}{A dataframe with grouped mz and RT values for a particular cluster.}
}
\value{
list of samples' intensities
}
\description{
Create intensity row for batch data
}
