% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_metadata}
\alias{create_metadata}
\title{Create a metadata row tibble with min, max and mean mz and RT values.}
\usage{
create_metadata(sample_grouped, sample_names)
}
\arguments{
\item{sample_grouped}{A dataframe with grouped mz and RT values for a particular cluster.}

\item{sample_names}{A list of sample names.}
}
\description{
Create a metadata row tibble with min, max and mean mz and RT values.
}
