% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{create_output}
\alias{create_output}
\title{Create a list containing 3 tibbles: metadata, intensities and RTs.}
\usage{
create_output(sample_grouped, sample_names)
}
\arguments{
\item{sample_grouped}{A dataframe with grouped mz and RT values for a particular cluster.}

\item{sample_names}{A list of sample names.}
}
\value{
A list containing 3 tibbles: metadata, intensities and RTs.
}
\description{
Create a list containing 3 tibbles: metadata, intensities and RTs.
}
