% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_plot}
\alias{draw_plot}
\title{Draw basic plot with optional label.}
\usage{
draw_plot(
  x = c(-1, 1),
  y = c(-1, 1),
  xlab = "",
  ylab = "",
  main = "",
  axes = FALSE,
  type = "n",
  label = NA,
  cex = 1.2
)
}
\arguments{
\item{x}{A numeric vector specifying the x-axis range.}

\item{y}{A numeric vector specifying the y-axis range.}

\item{xlab}{A character string for the x-axis label.}

\item{ylab}{A character string for the y-axis label.}

\item{main}{A character string for the main title of the plot.}

\item{axes}{A logical value indicating whether to draw the axes.}

\item{type}{A character string specifying the type of plot (default is "n" for no plotting).}

\item{label}{A character string for the text label to be placed at the center of the plot (default is NA, meaning no label).}

\item{cex}{A numeric value specifying the size of the text label (default is 1.2).}
}
\description{
This function creates a basic plot with specified x and y ranges, labels, and an optional text label at the center.
}
