% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_rt_correction_plot}
\alias{draw_rt_correction_plot}
\title{Draw retention time correction plot.}
\usage{
draw_rt_correction_plot(
  colors,
  extracted_features,
  corrected_features,
  rt_tol_relative
)
}
\arguments{
\item{colors}{A character vector of colors to use for plotting each sample. If NA, default colors are used.}

\item{extracted_features}{A list of data frames, where each data frame contains the original retention times and m/z values of features.}

\item{corrected_features}{A list of data frames, where each data frame contains the corrected retention times of features.}

\item{rt_tol_relative}{A numeric value representing the relative retention time tolerance.}
}
\description{
This function draws a plot showing the retention time correction for extracted features.
It plots the deviation of corrected retention times from the original retention times.
}
