% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_rt_normal_peaks}
\alias{draw_rt_normal_peaks}
\title{Draw retention time normal peaks.}
\usage{
draw_rt_normal_peaks(x, truth)
}
\arguments{
\item{x}{A numeric vector of retention time values.}

\item{truth}{A numeric vector containing the parameters for the normal peaks:
\itemize{
  \item mean of the peak
  \item standard deviation for the left side of the peak
  \item standard deviation for the right side of the peak
  \item scaling factor for the peak height
}}
}
\description{
This function draws the normal peaks for retention time data based on the provided parameters.
}
