% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{duplicate.row.remove}
\alias{duplicate.row.remove}
\title{Custom way of removing duplicate rows from a specifically formatted table.}
\usage{
duplicate.row.remove(features, tolerance = 1e-10)
}
\arguments{
\item{tolerance}{Tolerance to use for numeric comparisons.}

\item{new.table}{The table from which the duplicate rows should be removed. Needs at least 5 columns.
Columns 1, 2 and 5 have to be of numeric type.}
}
\value{
Returns the same table with duplicate rows removed.
}
\description{
Rows are considered as duplicate if the 1st, 2nd and 5th column are less than 1e-10 (tolerance) apart.
Only a single row in this `range` is kept from a group.
}
