% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{enrich_table_by_known_features}
\alias{enrich_table_by_known_features}
\title{Add entries from the known features table to the aligned table.}
\usage{
enrich_table_by_known_features(
  aligned,
  known_table,
  match_tol_ppm,
  mz_tol_relative,
  rt_tol_relative
)
}
\arguments{
\item{aligned}{A list object with three tibble tables: metadata, intensity, and rt.}

\item{known_table}{A table of known/previously detected peaks.}

\item{match_tol_ppm}{The ppm tolerance to match identified features to known metabolites/features.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
the tolerance level based on the data.}
}
\value{
Aligned table with known features.
}
\description{
Add entries from the known features table to the aligned table.
}
