% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{feature extraction}
\usage{
extract_features(
  cluster,
  filenames,
  min_pres,
  min_run,
  mz_tol,
  baseline_correct,
  baseline_correct_noise_percentile,
  intensity_weighted,
  min_bandwidth,
  max_bandwidth,
  sd_cut,
  sigma_ratio_lim,
  shape_model,
  peak_estim_method,
  component_eliminate,
  moment_power,
  BIC_factor
)
}
\arguments{
\item{cluster}{The number of CPU cores to be used}

\item{filenames}{The CDF file names.}

\item{min_pres}{This is a parameter of thr run filter, to be passed to the function proc.cdf().}

\item{min_run}{Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z 
to be considered a peak.}

\item{mz_tol}{m/z tolerance level for the grouping of data points. This value is expressed as the fraction of 
the m/z value. This value, multiplied by the m/z value, becomes the cutoff level. The recommended value is 
the machine's nominal accuracy level. Divide the ppm value by 1e6. For FTMS, 1e-5 is recommended.}

\item{baseline_correct}{After grouping the observations, the highest intensity in each group is found. If the highest is lower than 
this value, the entire group will be deleted. The default value is NA, in which case the program uses the 75th percentile of the 
height of the noise groups.}

\item{baseline_correct_noise_percentile}{The perenctile of signal strength of those EIC that don't pass the 
run filter, to be used as the baseline threshold of signal strength.}

\item{intensity_weighted}{Whether to use intensity to weight mass density estimation.}

\item{min_bandwidth}{The minimum bandwidth to use in the kernel smoother.}

\item{max_bandwidth}{The maximum bandwidth to use in the kernel smoother.}

\item{sd_cut}{A vector of two. Features with standard deviation outside the range defined by the two numbers 
are eliminated.}

\item{sigma_ratio_lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard 
deviation and the righ-standard deviation of the bi-Gaussian fuction used to fit the data.}

\item{shape_model}{The mathematical model for the shape of a peak. There are two choices - "bi-Gaussian" and 
"Gaussian". When the peaks are asymmetric, the bi-Gaussian is better. The default is "bi-Gaussian".}

\item{peak_estim_method}{The estimation method for the bi-Gaussian peak model. Two possible values: moment and EM.}

\item{component_eliminate}{In fitting mixture of bi-Gaussian (or Gaussian) model of an EIC, when a component accounts 
for a proportion of intensities less than this value, the component will be ignored.}

\item{moment_power}{The power parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture 
model in an EIC.}

\item{BIC_factor}{The factor that is multiplied on the number of parameters to modify the BIC criterion. 
If larger than 1, models with more peaks are penalized more.}
}
\description{
extract feature
}
\examples{
extract_features(cluster, filenames, min_pres, min_run, mz_tol, 0, 0.05, intensity_weighted, NA, NA, sd_cut, sigma_ratio_lim, "bi-Gaussian", "moment", 0.01, 1, 2.0)
}
