% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.step.hybrid.R
\name{extract_pattern_colnames}
\alias{extract_pattern_colnames}
\title{Extract column names matching a pattern.}
\usage{
extract_pattern_colnames(dataframe, pattern)
}
\arguments{
\item{dataframe}{A dataframe from which to extract column names.}

\item{pattern}{A character string containing the pattern to match in the column names.}
}
\value{
A character vector of column names that match the specified pattern.
}
\description{
This function extracts the column names from a dataframe that contain a specified pattern.
}
\keyword{internal}
