% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{fill_missing_values}
\alias{fill_missing_values}
\title{Fill missing values based on original retention times.}
\usage{
fill_missing_values(orig.feature, this.feature)
}
\arguments{
\item{orig.features}{Non-corrected feature table.}

\item{this.features}{Feature table with eventual missing values.}
}
\value{
Tibble Feature table with filles values.
}
\description{
Fill missing values based on original retention times.
}
