% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{filter_based_on_density}
\alias{filter_based_on_density}
\title{Subset data within lower and upper bound from density estimation}
\usage{
filter_based_on_density(sample, turns, index, i)
}
\arguments{
\item{sample}{A subset of the features_table.}

\item{turns}{A list of peaks and valleys positions.}

\item{index}{Whether it subsets on m/z [1] or RT [2] column.}

\item{i}{Iterates over the peaks in the turns list.}
}
\value{
Dataframe subsetted within lower and upper bound from density estimation.
}
\description{
Subset data within lower and upper bound from density estimation
}
