% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.match.R
\name{find.match}
\alias{find.match}
\title{Internal function: finding the best match between a set of detected features and a set of known features.}
\usage{
find.match(distances, max_distance)
}
\arguments{
\item{distances}{A matrix of distances.}

\item{max_distance}{A distance larger than which cannot be accepted as pairs.}
}
\value{
A binary matrix the same size as the input matrix, with matched position taking value 1, and all other positions taking value 0.
}
\description{
Given a small matrix of distances, find the best column-row pairing that minimize the sum of distances of the matched pairs.
}
