% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.turn.point.R
\name{find_local_maxima}
\alias{find_local_maxima}
\title{Compute local maxima turn points.}
\usage{
find_local_maxima(y, ties.method)
}
\arguments{
\item{y}{The y values of a curve in x-y plane.}

\item{ties.method}{specifies the method rank uses to break ties.}
}
\value{
boolean row with local maxima turn point.
}
\description{
Compute local maxima turn points.
}
