% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.match.R
\name{find_min_position}
\alias{find_min_position}
\title{Internal function: finding the position of the minimum value in a distance matrix.
Given a small matrix of distances, find the position of the minimum value.}
\usage{
find_min_position(distances)
}
\arguments{
\item{distances}{A matrix of distances.}
}
\value{
A vector of length 2 indicating the row and column of the minimum value.
}
\description{
Internal function: finding the position of the minimum value in a distance matrix.
Given a small matrix of distances, find the position of the minimum value.
}
