% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{find_mz_match}
\alias{find_mz_match}
\title{Compute matches between mz array and specific mass value with a tolerance.}
\usage{
find_mz_match(sample_mz, known_mz, match_tol_ppm = 5)
}
\arguments{
\item{sample_mz}{The mz array for which to compute the matching.}

\item{known_mz}{The mz value with which to match.}

\item{match_tol_ppm}{Matching tolerance in ppm.}
}
\value{
Indicies of m/z values within the tolerance of any known m/z.
}
\description{
Compute matches between mz array and specific mass value with a tolerance.
}
\examples{
find_mz_match(
 sample_mz = c(10, 20, 21),
 known_mz = 20
)
}
