% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{find_optima}
\alias{find_optima}
\title{Compute the kernel density estimation and find the peaks and valleys of a smooth curve.}
\usage{
find_optima(data, bandwidth)
}
\arguments{
\item{data}{A vector of m/z or RTs for a particular cluster.}

\item{bandwidth}{A bandwidth value for the KDE computation.}
}
\value{
A list of peaks and valleys positions.
}
\description{
Compute the kernel density estimation and find the peaks and valleys of a smooth curve.
}
