% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{get_custom_rt_tol}
\alias{get_custom_rt_tol}
\title{Compute custom chromatographic tolerance.}
\usage{
get_custom_rt_tol(use_observed_range, peak_rts, rt_range, aligned_features)
}
\arguments{
\item{use_observed_range}{bool Whether to use the observed chromatographic range for computation or not.}

\item{peak_rts}{data.frame Retention time cross table with all peak rts.}

\item{rt_range}{float Default chromatographic tolerance to use.}

\item{aligned_features}{data.frame Aligned feature table.}
}
\value{
vector Custom chromatographic tolerances to use for each feature.
}
\description{
Compute chromatographic tolerance for each feature. If `use_observed_range == TRUE`,
the whole range of retention times for all peaks is used to compute the tolerance,
otherwise `rt_range` is used for each feature.
}
