% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{get_mzrange_bound_indices}
\alias{get_mzrange_bound_indices}
\title{Compute range of valley indices which are in mz_tol range around aligned_feature_mass.}
\usage{
get_mzrange_bound_indices(aligned_feature_mass, mz, vlys_indices, mz_tol)
}
\arguments{
\item{aligned_feature_mass}{float Mz value of the aligned feature.}

\item{mz}{vector mz values of the features.}

\item{vlys_indices}{vector Indices of the valley points of mz clusters.}

\item{mz_tol}{float Tolerance to use to check if values are close.}
}
\value{
pair Index range (start, end).
}
\description{
Generate indices of valley points which are within the mz tolerance range around the aligned feature mass. If the aligned feature mass is smaller than the second valley point, the first two valley points are returned as indices.
Otherwise, the function checks which valley points are within the mz tolerance range and also includes the closest valley point above and below the aligned feature mass.
}
\keyword{internal}
