% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_workers}
\alias{get_num_workers}
\title{Get the number of available worker cores.}
\usage{
get_num_workers()
}
\value{
An integer representing the number of available worker cores.
}
\description{
This function determines the number of available worker cores, taking into account CRAN's limit on the number of cores.
}
