% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{get_rt_region_indices}
\alias{get_rt_region_indices}
\title{Get indices where rt in `features` is within `rt_tol` of `target_time`.}
\usage{
get_rt_region_indices(target_time, features, rt_tol)
}
\arguments{
\item{target_time}{float Target retention time region.}

\item{features}{tibble Feature table including `rt` column.}

\item{rt_tol}{float Retention time tolerance.}
}
\value{
vector Indices which are within `rt_tol` from `target_time` or
1 if `target_time` is NA.
}
\description{
Get indices where rt in `features` is within `rt_tol` of `target_time`.
}
