% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Get the sample name as basename of the file.}
\usage{
get_sample_name(filename)
}
\arguments{
\item{filename}{string Name of the file.}
}
\value{
string Sample name.
}
\description{
Get the sample name as basename of the file.
}
