% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Get the sample names as column sample_id in a profile table.}
\usage{
get_sample_names(profiles)
}
\arguments{
\item{profiles}{Feature tables.}
}
\value{
characted vector.
}
\description{
Get the sample names as column sample_id in a profile table.
}
