% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{get_single_occurrence_mask}
\alias{get_single_occurrence_mask}
\title{Get boolean mask for values that occur only once.}
\usage{
get_single_occurrence_mask(values)
}
\arguments{
\item{values}{vector Values for which to compute the mask.}
}
\value{
vector Boolean vector which is the mask of values occuring only once.
}
\description{
Uses the \link[base]{table} function to compute the occurrences and then
checks which values only occur a single time.
}
