% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{get_times_to_use}
\alias{get_times_to_use}
\title{Get retention time values to use}
\usage{
get_times_to_use(original_sample_rts, adjusted_sample_rts, cap = 2000)
}
\arguments{
\item{original_sample_rts}{vector Original feature retention time values before correction.}

\item{adjusted_sample_rts}{vector Feature retention time values after time correction.}

\item{cap}{int Maximum number of time values to return.}
}
\value{
Indices of retention time values to use.
}
\description{
Obtain retention time values which occur only once in both the original and the adjusted times.
This is a custom version of the unique or intersection function with rounding etc.
}
