% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive.bin.R
\name{increment_counter}
\alias{increment_counter}
\title{Increment pointers in a list.}
\usage{
increment_counter(pointers, that.n)
}
\arguments{
\item{pointers}{A list containing the pointers to be incremented.}

\item{that.n}{An integer value to increment the `prof.pointer`.}
}
\value{
The updated list of pointers.
}
\description{
This function increments the values of `prof.pointer`, `height.pointer`, and `curr.label` in the provided list of pointers.
}
\keyword{internal}
