% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{l2normalize}
\alias{l2normalize}
\title{Normalize vector so that sum(vec) = 1}
\usage{
l2normalize(x)
}
\arguments{
\item{x}{Data to normalize.}
}
\value{
Normalized data.
}
\description{
x / sum(x)
}
