% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_filter.R
\name{label_val_to_keep}
\alias{label_val_to_keep}
\title{This function labels the indices of values kept to perform further calculations}
\usage{
label_val_to_keep(min_run, timeline, min_pres, this_times, times)
}
\arguments{
\item{min_run}{Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z to be considered a peak.}

\item{min_pres}{Run filter parameter. The minimum proportion of presence in the time period for a series of signals grouped
by m/z to be considered a peak.}

\item{times.}{Retention times vector.}
}
\value{
to_keep.
}
\description{
This function labels the indices of values kept to perform further calculations
}
