% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lcms.R
\name{load.lcms.raw}
\alias{load.lcms.raw}
\title{Loading MS data from raw files.}
\usage{
load.lcms.raw(filename, chunk_size = 100)
}
\arguments{
\item{filename}{The raw file name.}
}
\value{
A tibble is returned with the following columns:
\itemize{
  \item mz - The vector of m/z values.
  \item rt - The vector of retention times.
  \item intensities - The vector of intensity values.
}
}
\description{
This is an internal function. It loads MS data from raw files into memory.
}
