% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lcms.R
\name{load.lcms}
\alias{load.lcms}
\title{Loading LC/MS data.}
\usage{
load.lcms(filename)
}
\arguments{
\item{filename}{The CDF file name.}
}
\value{
A tibble is returned with the following columns:
\itemize{
  \item mz - The vector of m/z values.
  \item rt - The vector of retention times.
  \item intensity - The vector of intensity values.
}
}
\description{
This is an internal function. It loads LC/MS data into memory.
}
