% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_noise.R
\name{load_data}
\alias{load_data}
\title{Load data either from cache or load raw file and detect peaks.}
\usage{
load_data(filename, cache, min_run, min_pres, mz_tol, intensity_weighted)
}
\description{
THIS FUNCTION IS DEPRECATED!
}
