% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.match.R
\name{mass.match}
\alias{mass.match}
\title{Compute matches between mz array and specific mass value with a tolerance.}
\usage{
mass.match(x, known.mz, match.tol.ppm = 5)
}
\arguments{
\item{x}{The mz array for which to compute the matching.}

\item{known.mz}{The mz value with which to match.}

\item{match.tol.ppm}{Matching tolerance in ppm.}
}
\value{
Binary vector, 1 indicating a match, 0 a mismatch.
}
\description{
Compute matches between mz array and specific mass value with a tolerance.
}
\examples{
mass.match(aligned.ftrs, known.table, match.tol.ppm)
}
