% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.turn.point.R
\name{msExtrema}
\alias{msExtrema}
\title{Compute maxima and minima turn points.}
\usage{
msExtrema(y)
}
\arguments{
\item{y}{The y values of a curve in x-y plane.}
}
\value{
boolean row with local maxima and minima turn points.
}
\description{
Compute maxima and minima turn points.
}
