% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{normix.bic}
\alias{normix.bic}
\title{Estimates parameters of a gaussian curve with normix and selects the best model using BIC criterion.}
\usage{
normix.bic(
  rt_profile,
  do.plot = FALSE,
  bw = c(15, 30, 60),
  eliminate = 0.05,
  max.iter = 50,
  BIC_factor = 2,
  aver_diff
)
}
\arguments{
\item{rt_profile}{Dataframe that stores RTs and intensities of features.}

\item{bw}{Bandwidth vector to use in the kernel smoother.}

\item{eliminate}{When a component accounts for a proportion of intensities less than this value, the component will be ignored.}

\item{max.iter}{Maximum number of iterations when executing the E step.}

\item{BIC_factor}{The factor that is multiplied on the number of parameters to modify the BIC criterion. If larger than 1,}

\item{aver_diff}{Average retention time difference across RTs of all features.}

\item{x}{Vector of RTs that lay in the same RT cluster.}

\item{y}{Intensities that belong to x.}
}
\description{
Returns named list containing the parameters of the best model and records of all models.
}
