% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{normix}
\alias{normix}
\title{Reevaluate parameters of chromatographic gaussian curves.}
\usage{
normix(rt_profile, pks, vlys, ignore = 0.1, max.iter = 50, aver_diff)
}
\arguments{
\item{pks}{A vector of sorted RT-peak values at which the kernel estimate was computed.}

\item{vlys}{A vector of sorted RT-valley values at which the kernel estimate was computed.}

\item{ignore}{In fitting mixture of bi-Gaussian (or Gaussian) model of an EIC, when a component accounts for a
proportion of intensities less than this value, the component will be ignored.}

\item{max.iter}{Maximum number of iterations when reevaluating gaussian curves.}

\item{aver_diff}{Average retention time difference across RTs of all features.}

\item{that.curve}{Dataframe that stores RTs and intensities of features.}
}
\value{
A list. The items are as follows going from first to last:
\itemize{
  \item miu - float - mean value of the gaussian curve
  \item sigma - float - standard deviation of the gaussian curve
  \item scale - float - estimated total signal strength (total area of the estimated normal curve)
}
}
\description{
Estimates Gaussian Mixture Model parameters with Expectation-Maximization (EM) algorithm.
Returns a matrix with columns: miu, sigma, scale, each row corresponds to a peak
}
