% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_normix_bic}
\alias{plot_normix_bic}
\title{Plot normalized mixture model with BIC.}
\usage{
plot_normix_bic(x, y, bw, aaa)
}
\arguments{
\item{x}{A numeric vector of data points on the x-axis.}

\item{y}{A numeric vector of data points on the y-axis.}

\item{bw}{The bandwidth used for the kernel density estimation.}

\item{aaa}{A matrix containing the parameters of the Gaussian mixture model components:
\itemize{
  \item mean of the Gaussian component
  \item standard deviation of the Gaussian component
  \item scaling factor for the Gaussian component
}}
}
\description{
This function plots the data points and the fitted Gaussian mixture model components with different bandwidths.
}
