% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_peak_summary}
\alias{plot_peak_summary}
\title{Plot peak summary.}
\usage{
plot_peak_summary(feature_groups, processed_features)
}
\arguments{
\item{feature_groups}{A list of data frames, where each data frame represents a group of features with m/z values.}

\item{processed_features}{A data frame containing processed feature information with columns "sd1", "sd2", and "area".}
}
\description{
This function plots a summary of peak characteristics, including m/z standard deviation, retention time standard deviation, and peak strength.
}
