% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_raw_profile_histogram}
\alias{plot_raw_profile_histogram}
\title{Plot raw profile histogram.}
\usage{
plot_raw_profile_histogram(
  raw.prof,
  min_pres,
  baseline.correct,
  baseline.correct.noise.percentile,
  mz_tol,
  new.prof
)
}
\arguments{
\item{raw.prof}{A list containing raw profile data, including height records and minimum count run.}

\item{min_pres}{A numeric value indicating the minimum presence threshold.}

\item{baseline.correct}{A numeric value for baseline correction. If NA, it will be computed automatically.}

\item{baseline.correct.noise.percentile}{A numeric value indicating the percentile of noise group heights for baseline correction.}

\item{mz_tol}{A numeric value for the mass-to-charge ratio tolerance.}

\item{new.prof}{A list containing new profile data, including height records, time range records, and m/z presence records.}
}
\description{
This function plots various histograms and density plots for a given raw profile, including noise groups, selected groups, retention time range distribution, and signal presence distribution.
}
