% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rt_profile}
\alias{plot_rt_profile}
\title{Plot retention time profile.}
\usage{
plot_rt_profile(rt_profile, bw, fit, m)
}
\arguments{
\item{rt_profile}{A data frame containing the retention time profile with columns "base_curve" and "intensity".}

\item{bw}{The bandwidth used for the kernel density estimation.}

\item{fit}{A matrix containing the fitted components for the retention time profile.}

\item{m}{A numeric vector of positions where vertical lines should be drawn.}
}
\description{
This function plots the retention time profile, including the base curve, intensity, and fitted components.
}
