% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{predict_mz_break_indices}
\alias{predict_mz_break_indices}
\title{Predict the indices for the valley points with low mass density.}
\usage{
predict_mz_break_indices(features, mz_orig_tol)
}
\arguments{
\item{features}{data.frame Data table with features for which to predict the indices.}

\item{mz_orig_tol}{float Mz tolerance to use as KDE bandwidth parameter.}
}
\value{
vector Predicted indices for valley points.
}
\description{
The density of mz values in the feature table is computed based on the tolerance.
The valleys or breaks of clusters in mz values are detected and a function
is approximated to predict the indices for the mass values which are the closest to those
valley points.
}
