% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_filter.R
\name{predict_smoothed_rt}
\alias{predict_smoothed_rt}
\title{Computes the smoothed retention times by using The Nadaraya-Watson kernel regression estimate function.}
\usage{
predict_smoothed_rt(min_run, times)
}
\arguments{
\item{min_run}{Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z to be considered a peak.}

\item{times.}{Retention times vector.}
}
\value{
predicted rt.
}
\description{
Computes the smoothed retention times by using The Nadaraya-Watson kernel regression estimate function.
}
