% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{prep.uv}
\alias{prep.uv}
\title{This function prepares the parameters required for latter compuation. u, v, and sum of x (original authors' comment).}
\usage{
prep.uv(x, t, a)
}
\description{
This function prepares the parameters required for latter compuation. u, v, and sum of x (original authors' comment).
}
