% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{prep_uv}
\alias{prep_uv}
\title{Compute 'u' and 'v' parameters of bi-gaussian chromatographic peak shape}
\usage{
prep_uv(x, t, a)
}
\arguments{
\item{x}{A vector of numerical values (intensities).}

\item{t}{A vector of numerical values (rt).}

\item{a}{A vector of peak summits.}
}
\description{
This function prepares the parameters required for latter computation.
u, v, and sum of x (original authors' comment).
}
\keyword{internal}
