% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{preprocess_bandwidth}
\alias{preprocess_bandwidth}
\title{Initialize minimum and maximum bandwidth values if none given. Ensure that minimum bandwidth is lower that maximum, else set minimum to 1/4 of maximum value.}
\usage{
preprocess_bandwidth(min_bandwidth, max_bandwidth, profile)
}
\arguments{
\item{min_bandwidth}{The minimum bandwidth to use in the kernel smoother.}

\item{max_bandwidth}{The maximum bandwidth to use in the kernel smoother.}

\item{profile}{Profile table with shape number-of-features*4. The table contains following columns:
\itemize{
  \item mz - float - mass-to-charge ratio of feature
  \item rt - float - retention time of features
  \item intensity - float - intensity of features
  \item group_number - integer - group number assigned to each feature based on their rt similarity
}}
}
\value{
Returns a list object with the following objects in it:
\itemize{
  \item min_bandwidth - float - Minimum bandwidth.
  \item max_bandwidth - float - Maximum bandwidth
}
}
\description{
Initialize minimum and maximum bandwidth values if none given. Ensure that minimum bandwidth is lower that maximum, else set minimum to 1/4 of maximum value.
}
