% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{preprocess_profile}
\alias{preprocess_profile}
\title{Convert input matrix to a dataframe with column names (see source code for the names).}
\usage{
preprocess_profile(profile)
}
\arguments{
\item{profile}{Profile table with shape number-of-features*4. The table contains following columns:
\itemize{
  \item float - mass-to-charge ratio of feature
  \item float - retention time of features
  \item float - intensity of features
  \item integer - group number assigned to each feature based on their rt similarity
}}
}
\value{
Returns a dataframe with shape number-of-features*4. The columns are as follows:
\itemize{
  \item mz - float - mass-to-charge ratio of feature
  \item rt - float - retention time of features
  \item intensity - float - intensity of features
  \item group_number - integer - group number assigned to each feature based on their rt similarity
}
}
\description{
Convert input matrix to a dataframe with column names (see source code for the names).
}
