% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lcms.R
\name{process_chunk}
\alias{process_chunk}
\title{Process a chunk of spectra and return a list of m/z, rt, and intensities.
#' @param spectra A list of spectra, where each spectrum is a list containing m/z and intensity values.}
\usage{
process_chunk(spectra, start_times)
}
\arguments{
\item{start_times}{A vector of start times corresponding to each spectrum.}
}
\value{
A list of lists, where each inner list contains m/z, rt, and intensities for a spectrum.
}
\description{
Process a chunk of spectra and return a list of m/z, rt, and intensities.
#' @param spectra A list of spectra, where each spectrum is a list containing m/z and intensity values.
}
