% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{rev_cum_sum}
\alias{rev_cum_sum}
\title{Computes vector of cumulative sums on reversed input. Returns cumulative sum vector going from the sum of all elements to one.}
\usage{
rev_cum_sum(x)
}
\arguments{
\item{x}{float - vector of numerical values}
}
\value{
Returns a vector
}
\description{
Computes vector of cumulative sums on reversed input. Returns cumulative sum vector going from the sum of all elements to one.
}
