% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{solve.a}
\alias{solve.a}
\title{This function solves the value of a using the x, t, a from the previous step, and sigma.1, and sigma.2 (original authors' comment).}
\usage{
\method{solve}{a}(x, t, a, sigma.1, sigma.2)
}
\description{
This function solves the value of a using the x, t, a from the previous step, and sigma.1, and sigma.2 (original authors' comment).
}
