% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{solve.sigma}
\alias{solve.sigma}
\title{This function takes the value intensity level x, retention time t and assumed breaking point a, calculates the square estimated of sigma.1 and sigma.2 (original authors' comment).}
\usage{
\method{solve}{sigma}(x, t, a)
}
\description{
This function takes the value intensity level x, retention time t and assumed breaking point a, calculates the square estimated of sigma.1 and sigma.2 (original authors' comment).
}
