% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{solve_a}
\alias{solve_a}
\title{Compute 'a' parameter of bi-gaussian chromatographic peak shape}
\usage{
solve_a(x, t, a, sigma.1, sigma.2)
}
\arguments{
\item{x}{A vector of numerical values (intensities).}

\item{t}{A vector of numerical values (rt).}

\item{a}{A vector of peak summits.}

\item{sigma.1}{left standard deviation of the gaussian curve}

\item{sigma.2}{right standard deviation of the gaussian curve}
}
\description{
This function solves peak summit using the x, t, peak summit from the previous step, 
and sigma.1, and sigma.2 (original authors' comment).
}
