% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{solve_sigma}
\alias{solve_sigma}
\title{Solve sigma}
\usage{
solve_sigma(x, t, a)
}
\arguments{
\item{x}{A vector of numerical values (intensities).}

\item{t}{A vector of numerical values (rt).}

\item{a}{A vector of peak summits.}
}
\value{
A vector of:
\itemize{
  \item standard deviation at the left side of the gaussian curve
  \item standard deviation at the right side of the gaussian curve
}
}
\description{
Calculates the square estimated of left and right standard deviations.
}
