% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{span}
\alias{span}
\title{Calculate the span of a numeric vector.}
\usage{
span(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the span of the vector.
}
\description{
This function calculates the span (range) of a numeric vector, ignoring NA values.
}
