% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tolerance_plot}
\alias{tolerance_plot}
\title{Plot tolerance level.}
\usage{
tolerance_plot(x, y, exp_y, selected, main)
}
\arguments{
\item{x}{A numeric vector of x values.}

\item{y}{A numeric vector of y values.}

\item{exp_y}{A numeric vector of expected y values.}

\item{selected}{An integer index indicating the selected point for drawing a line.}

\item{main}{A character string for the main title of the plot.}
}
\description{
This function plots the tolerance level for a given set of x and y values, along with expected y values and a selected index.
}
