% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{validate_contents}
\alias{validate_contents}
\title{Validates if the data is present in more than "min_occurence" of samples.}
\usage{
validate_contents(samples, min_occurrence)
}
\arguments{
\item{samples}{A subset of the features_table.}

\item{min_occurrence}{A minimal number of profiles a feature has to be present in.}
}
\value{
boolean value whether it is TRUE or FALSE.
}
\description{
Validates if the data is present in more than "min_occurence" of samples.
}
