% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate that provided inputs match expected, exit execution otherwise.}
\usage{
validate_inputs(shape.model, estim.method)
}
\arguments{
\item{shape.model}{The mathematical model for the shape of a peak. There are two choices - "bi-Gaussian" and "Gaussian".
When the peaks are asymmetric, the bi-Gaussian is better. The default is "bi-Gaussian".}

\item{estim.method}{The estimation method for the bi-Gaussian peak model. Two possible values: moment and EM.}
}
\description{
Validate that provided inputs match expected, exit execution otherwise.
}
