% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{validate_model_method_input}
\alias{validate_model_method_input}
\title{Validate that provided model and method inputs match expected, exit execution otherwise.}
\usage{
validate_model_method_input(shape_model, peak_estim_method)
}
\arguments{
\item{shape_model}{The mathematical model for the shape of a peak. There are two choices - "bi-Gaussian" and "Gaussian".
When the peaks are asymmetric, the bi-Gaussian is better. The default is "bi-Gaussian".}

\item{peak_estim_method}{The estimation method for the bi-Gaussian peak model. Two possible values: moment and EM.}
}
\description{
Validate that provided model and method inputs match expected, exit execution otherwise.
}
