% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_mu_sc_std}
\alias{compute_mu_sc_std}
\title{Compute interpolated retention time, its standard deviation, and intensity values,.}
\usage{
compute_mu_sc_std(features, aver_diff)
}
\arguments{
\item{features}{tibble Features with `rt` and `intensities` columns.}

\item{aver_diff}{float Average retention time difference.}
}
\value{

}
\description{
Compute interpolated retention time, its standard deviation, and intensity values,.
}
