% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.weaker.R
\name{compute_pks_vlys_rt}
\alias{compute_pks_vlys_rt}
\title{Compute peaks and valleys which have at least `recover_min_count` peaks.}
\usage{
compute_pks_vlys_rt(features, times, bandwidth, target_rt, recover_min_count)
}
\arguments{
\item{features}{tibble Features with `mz`, `rt` and `intensities`.}

\item{times}{vector Retention time values from the raw data file.}

\item{bandwidth}{float Bandwidth to use in smoothing.}

\item{target_rt}{float Retention time at which to recover the intensity.}

\item{recover_min_count}{int Minimum number of peaks required in the area to recover the signal.}
}
\value{

}
\description{
Compute peaks and valleys which have at least `recover_min_count` peaks.
}
