% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{compute_scale}
\alias{compute_scale}
\title{Estimate total signal strength (total area of the estimated normal curve).}
\usage{
compute_scale(y, d)
}
\arguments{
\item{y}{- float - a vector of intensities.}

\item{d}{- float - a vector of \emph{y} values in a gaussian curve.}

\item{scale}{- float - a vector of scaled intensity values.}
}
\description{
Estimate total signal strength (total area of the estimated normal curve).
}
