% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.cdf.R
\name{load_data}
\alias{load_data}
\title{Load data either from cache or load raw file and detect peaks.}
\usage{
load_data(
  filename,
  cache,
  min_elution_length,
  min_presence,
  mz_tol,
  baseline_correct,
  intensity_weighted
)
}
\description{
Load data either from cache or load raw file and detect peaks.
}
