% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.cdf.R
\name{load_data}
\alias{load_data}
\title{Load data either from cache or load raw file and detect peaks.}
\usage{
load_data(
  filename,
  cache,
  min.run,
  min.pres,
  tol,
  baseline.correct,
  intensity.weighted
)
}
\description{
Load data either from cache or load raw file and detect peaks.
}
