% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{normix}
\alias{normix}
\title{Reevaluate parameters of rtomatographic gaussian curves.}
\usage{
normix(that.curve, pks, vlys, ignore = 0.1, max.iter = 50, aver_diff)
}
\arguments{
\item{that.curve}{Dataframe that stores RTs and intensities of features.}

\item{pks}{A vector of sorted RT-peak values at which the kernel estimate was computed.}

\item{vlys}{A vector of sorted RT-valley values at which the kernel estimate was computed.}

\item{ignore}{In fitting mixture of bi-Gaussian (or Gaussian) model of an EIC, when a component accounts for a
proportion of intensities less than this value, the component will be ignored.}

\item{max.iter}{Maximum number of iterations when reevaluating gaussian curves.}

\item{aver_diff}{Average retention time difference across RTs of all features.}
}
\description{
Reevaluate parameters of rtomatographic gaussian curves.
}
