% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_arrow}
\alias{write_arrow}
\title{Function to save table with metadata to parquet file}
\usage{
write_arrow(data_frame, path, metadata = NA)
}
\description{
Function to save table with metadata to parquet file
}
